+++ R/parser.R
@@ -206,9 +206,8 @@
 #' @return Named list.
 #' @export
 parse_multipart <- function(req, ...) {
-  on.exit(req$rook.input$rewind())
-  body <- req$rook.input$read()
-  if (identical(body, raw())) {
+  body <- req$body
+  if (is.null(body) || identical(body, raw())) {
     return(list())
   }
 
@@ -279,9 +278,8 @@ parse_multipart <- function(req, ...) {
 #' @return Named list
 #' @export
 parse_form_urlencoded <- function(req, ...) {
-  on.exit(req$rook.input$rewind())
-  body <- req$rook.input$read()
-  if (identical(body, raw())) {
+  body <- req$body
+  if (is.null(body) || identical(body, raw())) {
     return(list())
   }
 
@@ -332,9 +330,8 @@ parse_form_urlencoded <- function(req, ...) {
 #' @return Named list
 #' @export
 parse_json <- function(req, ...) {
-  on.exit(req$rook.input$rewind())
-  body <- req$rook.input$read()
-  if (identical(body, raw())) {
+  body <- req$body
+  if (is.null(body) || identical(body, raw())) {
     return(list())
   }